Rizkidwi

See merge request group2073748/hotel!15

+++ src/main/java/com/jdt13/hotel/dto/KamarResponse.java
@@ -2,12 +2,14 @@ package com.jdt13.hotel.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class KamarResponse {
 
     private Integer id;
     private Integer noKamar;
-    private Double harga;
+    private BigDecimal harga;
     private String kategori;
     private String deskripsi;
 }
